scripts/requirements.txt
# Python dependencies for Parquet conversion and Azure upload

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0

# Azure SDK
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Optional: For advanced analytics
numpy>=1.24.0